+++ platzky/db/graph_ql_db.py
@@ -3,14 +3,13 @@
 # TODO: Rename file, extract to another library, remove gql and aiohttp from dependencies
 
 import threading
-from typing import Any
 
 from gql import Client, gql
 from gql.transport.aiohttp import AIOHTTPTransport
 from pydantic import Field
 
 from platzky.db.db import DB, DBConfig
-from platzky.db.exceptions import NotFoundError
+from platzky.db.graphql_blog_storage import GraphQLBlogStorage
 from platzky.db.graphql_plugin_config_repository import GraphQLPluginConfigRepository
 from platzky.models import MenuItem, Page, Post
 from platzky.plugin.plugin_config import PluginConfigBase
@@ -44,107 +43,6 @@ def db_from_config(config: GraphQlDbConfig) -> "GraphQL":
     return GraphQL(config.endpoint, config.token)
 
 
-def _standardize_comment(
-    comment: dict[str, Any],
-) -> dict[str, Any]:
-    """Standardize comment data structure from GraphQL response.
-
-    Args:
-        comment: Raw comment data from GraphQL response
-
-    Returns:
-        Standardized comment dictionary
-    """
-    return {
-        "author": comment["author"],
-        "comment": comment["comment"],
-        "date": comment["createdAt"],
-    }
-
-
-def _standardize_post(post: dict[str, Any]) -> dict[str, Any]:
-    """Standardize post data structure from GraphQL response.
-
-    Args:
-        post: Raw post data from GraphQL response
-
-    Returns:
-        Standardized post dictionary
-    """
-    return {
-        "author": post["author"]["name"],
-        "slug": post["slug"],
-        "title": post["title"],
-        "excerpt": post["excerpt"],
-        "contentInMarkdown": post["contentInRichText"]["html"],
-        "comments": [_standardize_comment(comment) for comment in post["comments"]],
-        "tags": post["tags"],
-        "language": post["language"],
-        "coverImage": {
-            "url": (post.get("coverImage") or {}).get("image", {}).get("url", ""),
-        },
-        "date": post["date"],
-        "css": post.get("css") or "",
-    }
-
-
-def _standardize_page(page: dict[str, Any]) -> dict[str, Any]:
-    """Standardize page data structure from GraphQL response.
-
-    Pages have fewer required fields than posts in the GraphQL schema.
-    This function provides sensible defaults for missing Post fields.
-
-    Args:
-        page: Raw page data from GraphQL response
-
-    Returns:
-        Standardized page dictionary compatible with Page model
-    """
-    return {
-        "author": page.get("author", ""),
-        "slug": page.get("slug", ""),
-        "title": page["title"],
-        "excerpt": page.get("excerpt", ""),
-        "contentInMarkdown": page["contentInMarkdown"],
-        "comments": [],
-        "tags": page.get("tags", []),
-        "language": page.get("language", "en"),
-        "coverImage": {
-            "url": (page.get("coverImage") or {}).get("url", ""),
-        },
-        "date": page.get("date"),
-        "css": page.get("css") or "",
-    }
-
-
-def _standardize_post_by_tag(post: dict[str, Any]) -> dict[str, Any]:
-    """Standardize post data from get_posts_by_tag GraphQL response.
-
-    Posts returned by tag query have fewer fields than full posts.
-    This function provides sensible defaults for missing Post fields.
-
-    Args:
-        post: Raw post data from GraphQL get_posts_by_tag response
-
-    Returns:
-        Standardized post dictionary compatible with Post model
-    """
-    return {
-        "author": post.get("author", ""),
-        "slug": post["slug"],
-        "title": post["title"],
-        "excerpt": post["excerpt"],
-        "contentInMarkdown": post.get("contentInMarkdown", ""),
-        "comments": [],
-        "tags": post["tags"],
-        "language": post.get("language", "en"),
-        "coverImage": {
-            "url": (post.get("coverImage") or {}).get("image", {}).get("url", ""),
-        },
-        "date": post["date"],
-    }
-
-
 class GraphQL(DB):
     """GraphQL database implementation for CMS integration."""
 
@@ -160,6 +58,7 @@ class GraphQL(DB):
         self._endpoint = endpoint
         self._headers = {"Authorization": "bearer " + token}
         self._local = threading.local()
+        self._blog_storage = GraphQLBlogStorage(endpoint, token)
         self._plugins_repository = GraphQLPluginConfigRepository(endpoint, token)
         super().__init__()
 
@@ -191,39 +90,7 @@ class GraphQL(DB):
         Returns:
             List of Post objects
         """
-        all_posts = gql("""
-            query MyQuery($lang: Lang!) {
-              posts(where: {language: $lang},  orderBy: date_DESC, stage: PUBLISHED){
-                createdAt
-                author {
-                    name
-                }
-                contentInRichText {
-                    html
-                    }
-                comments {
-                  comment
-                  author
-                  createdAt
-                  }
-                date
-                title
-                excerpt
-                slug
-                tags
-                language
-                coverImage {
-                  alternateText
-                  image {
-                    url
-                  }
-                }
-              }
-            }
-            """)
-        raw_ql_posts = self.client.execute(all_posts, variable_values={"lang": lang})["posts"]
-
-        return [Post.model_validate(_standardize_post(post)) for post in raw_ql_posts]
+        return self._blog_storage.posts.get_all(lang)
 
     def get_menu_items_in_lang(self, lang: str) -> list[MenuItem]:
         """Retrieve menu items for a specific language.
@@ -257,44 +124,8 @@ class GraphQL(DB):
         Raises:
             NotFoundError: If no post exists for the given slug.
         """
-        post = gql("""
-            query MyQuery($slug: String!) {
-              post(where: {slug: $slug}, stage: PUBLISHED) {
-                date
-                language
-                title
-                slug
-                author {
-                    name
-                }
-                contentInRichText {
-                  markdown
-                  html
-                }
-                excerpt
-                tags
-                css
-                coverImage {
-                  alternateText
-                  image {
-                    url
-                  }
-                }
-                comments {
-                    author
-                    comment
-                    date: createdAt
-                }
-              }
-            }
-            """)
+        return self._blog_storage.posts.get(slug)
 
-        post_raw = self.client.execute(post, variable_values={"slug": slug})["post"]
-        if post_raw is None:
-            raise NotFoundError(f"Post not found: {slug}")
-        return Post.model_validate(_standardize_post(post_raw))
-
-    # TODO: Cleanup page logic of internationalization (now it depends on translation of slugs)
     def get_page(self, slug: str) -> Page:
         """Retrieve a page by its slug.
 
@@ -307,24 +138,7 @@ class GraphQL(DB):
         Raises:
             NotFoundError: If no page exists for the given slug.
         """
-        page_query = gql("""
-            query MyQuery ($slug: String!){
-              page(where: {slug: $slug}, stage: PUBLISHED) {
-                slug
-                title
-                contentInMarkdown
-                css
-                coverImage
-                {
-                    url
-                }
-              }
-            }
-            """)
-        page_raw = self.client.execute(page_query, variable_values={"slug": slug})["page"]
-        if page_raw is None:
-            raise NotFoundError(f"Page not found: {slug}")
-        return Page.model_validate(_standardize_page(page_raw))
+        return self._blog_storage.pages.get(slug)
 
     def get_posts_by_tag(self, tag: str, lang: str) -> list[Post]:
         """Retrieve posts filtered by tag and language.
@@ -336,25 +150,7 @@ class GraphQL(DB):
         Returns:
             List of Post objects
         """
-        post = gql("""
-            query MyQuery ($tag: String!, $lang: Lang!){
-              posts(where: {tags_contains_some: [$tag], language: $lang}, stage: PUBLISHED) {
-                    tags
-                    title
-                    slug
-                    excerpt
-                    date
-                    coverImage {
-                      alternateText
-                      image {
-                        url
-                      }
-                    }
-              }
-            }
-            """)
-        raw_posts = self.client.execute(post, variable_values={"tag": tag, "lang": lang})["posts"]
-        return [Post.model_validate(_standardize_post_by_tag(p)) for p in raw_posts]
+        return self._blog_storage.posts.get_by_tag(tag, lang)
 
     def add_comment(self, author_name: str, comment: str, post_slug: str) -> None:
         """Add a new comment to a post.
@@ -364,27 +160,7 @@ class GraphQL(DB):
             comment: Comment text content
             post_slug: URL-friendly identifier of the post
         """
-        add_comment = gql("""
-            mutation MyMutation($author: String!, $comment: String!, $slug: String!) {
-                createComment(
-                    data: {
-                        author: $author,
-                        comment: $comment,
-                        post: {connect: {slug: $slug}}
-                    }
-                ) {
-                    id
-                }
-            }
-            """)
-        self.client.execute(
-            add_comment,
-            variable_values={
-                "author": author_name,
-                "comment": comment,
-                "slug": post_slug,
-            },
-        )
+        self._blog_storage.posts.add_comment(author_name, comment, post_slug)
 
     def get_font(self) -> str:
         """Get the font configuration for the application.
